Collect interface methods with default implementations
A class implementing an interface with default method implementations
will not necessarily override those methods; in such cases, the methods
are currently "invisible" to DynamicProxy and thus won't get proxied.

This can be solved by an extra pass over all interfaces to collect such
methods, as is done in this commit.

We should next check whether the same work can be performed in the pre-
existing infrastructure, so that we don't need to iterate over all
interfaces more than once.