Need to collect default interface methods extra
Because default interface methods do not need to be implemented on class
types inheriting the interface, those methods can be absent from the
class and `ClassMembersCollector` will not see them; thus they don't get
proxied.

This commit adds an extra collector for interface methods that are
implemented in an interface (and not the proxied class).

While this solution works, it's not ideal from a performance standpoint,
as it adds an extra pass over all interfaces; we should check whether
the existing infrastructure can be adapted to do the same work, so that
only a single iteration over all interfaces would be required.